Moving limits interface to the tripperware package
